+++ frontend/src/pages/BillPage/components/BillCardList.tsx
@@ -1,26 +1,23 @@
-import { useState } from "react";
 import BillCard from "@/components/BillCard";
-import Pagination from "@/components/Pagination";
 import { sampleBills } from "@/mockdata/bills";
 
-export const BillCardList = () => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const itemsPerPage = 2;
-  const totalPages = Math.ceil(sampleBills.length / itemsPerPage);
+interface BillCardListProps {
+  currentPage: number;
+  itemsPerPage: number;
+}
 
+export const BillCardList = ({
+  currentPage,
+  itemsPerPage,
+}: BillCardListProps) => {
   const startIndex = (currentPage - 1) * itemsPerPage;
   const endIndex = startIndex + itemsPerPage;
   const currentBills = sampleBills.slice(startIndex, endIndex);
 
-  const handlePageChange = (page: number) => {
-    setCurrentPage(page);
-  };
-
   return (
     <>
-      {/* 카드 그리드 */}
       <div className="flex justify-center">
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 max-w-7xl">
+        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 max-w-[1200px]">
           {currentBills.map((bill) => (
             <BillCard
               key={bill.id}
@@ -35,12 +32,6 @@ export const BillCardList = () => {
           ))}
         </div>
       </div>
-
-      <Pagination
-        currentPage={currentPage}
-        totalPages={totalPages}
-        onPageChange={handlePageChange}
-      />
     </>
   );
 };
